Added manager/qs methods for getting other events in a given event's family, events with relatives satisfying a given queryset, and events with occurrences.
The pinnacle is the useful "highest_having_occurrences()" which filters the qs to show only events that have occurrences, but that also have no ancestors with occurrences.

Events.objects.highest_having_occurrences() is a good first blush at 'The List Of Events', since it is the longest list of events whose descendants' occurrences will cover the entire set of occurrences with no repetitions.